% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avg_table_country.R
\name{avg_table_country}
\alias{avg_table_country}
\title{Average democracy data by country}
\usage{
avg_table_country(dat)
}
\arguments{
\item{dat}{A data frame containing democracy data.}
}
\value{
A tibble with one row per country and average values for democracy and representation variables.
}
\description{
Summarizes democracy and population representation data by country.
}
