% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_required_cols.R
\name{check_required_cols}
\alias{check_required_cols}
\title{Check required columns}
\usage{
check_required_cols(dat, cols)
}
\arguments{
\item{dat}{A data frame.}

\item{cols}{A character vector of required column names.}
}
\value{
The original data frame if all required columns are present.
}
\description{
Checks whether a data frame contains all required columns.
}
\keyword{internal}
