% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/democracy_index_table.R
\name{democracy_index_table}
\alias{democracy_index_table}
\title{Create a Democracy Index table}
\usage{
democracy_index_table(democracy_index = load_democracy_index())
}
\arguments{
\item{democracy_index}{A data frame containing Democracy Index data.
Defaults to \code{load_democracy_index()}.}
}
\value{
A formatted \code{knitr::kable} table.
}
\description{
Creates a formatted table of countries with the highest Democracy Index scores.
}
\examples{
democracy_index_table()
democracy_index_table(democracy_index = load_democracy_index())
}
