% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demr-package.R
\docType{package}
\name{demr-package}
\alias{demr}
\alias{demr-package}
\title{demr: Geopolitical Structures Analysis}
\description{
Assists analysis on Geopolitical Structures.
}
\author{
\strong{Maintainer}: Rex Nelson \email{rnelso33@calpoly.edu}

Authors:
\itemize{
  \item Rex Nelson \email{rnelso33@calpoly.edu}
  \item Grant Cowles \email{gycowles@calpoly.edu}
  \item Talan Williams \email{twilli93@calpoly.edu}
}

}
\keyword{internal}
