% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_democracy_index}
\alias{load_democracy_index}
\title{Load Democracy Index data}
\usage{
load_democracy_index()
}
\value{
A tibble containing Democracy Index scores by country and year.
}
\description{
Loads the Economist Intelligence Unit Democracy Index data used in the demr package.
}
