% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_by_cont.R
\name{plot_by_cont}
\alias{plot_by_cont}
\title{Plot democracy score by representation type for one region}
\usage{
plot_by_cont(dat, continent)
}
\arguments{
\item{dat}{A data frame containing democracy data.}

\item{continent}{A character string giving the region to filter to.}
}
\value{
A ggplot object.
}
\description{
Creates a scatterplot of population proportion against spatial democracy score
for a selected region.
}
