% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_demscore.R
\name{plot_demscore}
\alias{plot_demscore}
\title{This is the first function to graph violin plots of democracy score by level of representation.}
\usage{
plot_demscore(level)
}
\arguments{
\item{level}{A string input to specify the level of representation}
}
\value{
A violin plot organized by continent
}
\description{
This is the first function to graph violin plots of democracy score by level of representation.
}
