% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data}
\alias{load_data}
\title{This is the first function to load our data.}
\usage{
load_data()
}
\value{
A data frame from dem_data.csv
}
\description{
This is the first function to load our data.
}
